Bug 848491 - Update Skia to r8495
UltraBlame original commit: b3e5868f25401ac2529fec6f3aaeaec6dc9e40b3

+++ gfx/skia/src/core/SkBitmap_scroll.cpp
@@ -82,7 +82,7 @@ bool SkBitmap::scrollRect(const SkIRect* subset, int dx, int dy,
 
     char*       dst = (char*)this->getPixels();
     const char* src = dst;
-    int         rowBytes = this->rowBytes();    
+    int         rowBytes = (int)this->rowBytes();    
 
     if (dy <= 0) {
         src -= dy * rowBytes;
